move code component
