+++ components/code/code-line/code-line.spec.tsx
@@ -1,0 +1,6 @@
+import { render, screen } from '@testing-library/react'
+import { CodeLine } from '.'
+
+test('renders CodeLine component', () => {
+    render(<CodeLine line="" />)
+})
