+++ components/code/code-line/code-line.stories.tsx
@@ -1,0 +1,14 @@
+import { CodeLine } from './code-line'
+
+import type { Meta, StoryObj } from '@storybook/react'
+
+type CodeLineType = typeof CodeLine
+
+export default {
+    title: 'CodeLine',
+    component: CodeLine,
+} satisfies Meta<CodeLineType>
+
+export const Default: StoryObj<CodeLineType> = {
+    render: () => <CodeLine line="" />,
+}
