+++ components/code/code.tsx
@@ -1,57 +1,10 @@
 'use client'
 
-import { useEffect, useRef } from 'react'
-
 import { useCode } from '@/components/hooks/use-code'
-import { useCursorIndex, useLineNumber } from '@/components/hooks/use-type'
-import { Caret } from '@/components/ui/caret'
+import { useLineNumber } from '@/components/hooks/use-type'
 import { Window } from '@/components/ui/window'
 
-import { extractLeadingWhitespace, highlightCodeAgain, scrollWithInScreen } from '@/lib/utils'
-
-const CodeTargetLine = ({ line }: { line: string }) => {
-    const { cursorIndex } = useCursorIndex(line)
-    const caretRef = useRef<HTMLSpanElement | null>(null)
-
-    useEffect(() => {
-        scrollWithInScreen(caretRef)
-    }, [])
-
-    const prefix = extractLeadingWhitespace(line)
-
-    return (
-        <code className="w-full whitespace-pre-wrap break-words">
-            <span>{line.slice(0, cursorIndex + prefix.length)}</span>
-            <Caret ref={caretRef} />
-            <span>{line.slice(cursorIndex + prefix.length)}</span>
-        </code>
-    )
-}
-
-const CodeLine = ({ line }: { line: string }) => {
-    const lineRef = useRef<HTMLElement | null>(null)
-
-    useEffect(() => {
-        highlightCodeAgain(lineRef)
-    }, [line])
-
-    return (
-        // tailwindcssのプロパティが効かないのでstyleで指定
-        <code
-            ref={lineRef}
-            style={{
-                padding: 0,
-                margin: 0,
-                backgroundColor: 'transparent',
-                whiteSpace: 'pre-wrap',
-                overflowWrap: 'break-word',
-            }}
-            className="typescript"
-        >
-            {line}
-        </code>
-    )
-}
+import { CodeLine, CodeTargetLine } from './code-line'
 
 const CodeLines = () => {
     const { code } = useCode()
